credrank: dd "bonus minting" in the Graph
This commit adds new logic to the bonus minting module so that we can
embed bonus minting directly in to the graph. That is to say, if we mint
bonus Cred for a particular recipient, there will be an explicit Graph
node that mints Cred and flows it to that recipient.

This is intended as a way to add bonus minting to CredRank, without
needing to special-case bonus minting outside of the usual Graph
abstraction. We considered a few approaches here, including embedding
complicated graph gadgets without minting any additional Cred, but this
approach is by far easier to implement.

The downside to this approach is that the amount of bonus-minted Cred
will not exactly match the intended specification. For example, if we
have a single bonus account and we want it to get 10% of the total Cred
(after bonus minting), we would set weight = 0.1111111. In practice,
using SC's own instance as a reference case, it winds up getting
8.6% of the Cred. It's impossible to get exact correctness using this
approach because of nuances in how much raw (PageRank) transition
probability accumulates at nodes based on the path between the seed node
and the recipient.

We could continue getting exact results (as in TimelineCred) if we
entirely special-case the bonus minting, rather than using the usual
Cred computation pipeline. However, I think it's preferrable to keep the
core algorithm as consistent as possible. I also see this kind of bonus
minting as a temporary solution to the general question of "how should
projects value their dependences". Therefore, I don't mind if it has a
few rough edges.

Note: This commit doesn't change behavior for TimelineCred at all. Since
we're planning to deprecate it, no sense upgrading it.

Test plan:
Run the `credrank` command on an instance that has dependencies
specified. You'll see that the dependency now gets Cred, and it's
generally similar to the intended minting amount. Also, unit tests
added.