Implement persist method on LedgerManager
The persist method writes the local ledger changes back to the ledger storage while
ensuring that there are no conflicts or inconsistencies in the ledger history as well
as minimizing the potential for remote changes to get overwritten due to race conditions.

This is not super airtight or ready to be used my many people frequently updating the
ledger at the same time, but it should cover the majority of sync conflicts that might
arise.

Test Plan: Unit tests + test with github storage backend once that's implemented